Move ControlsMarker and FunctionalMarker into a separate module, add documentation
